Refactor grid plugin structure and update relative file imports
