+++ src/plugins/web_view_grid/FNAbviewgridEditor.js
@@ -3,20 +3,19 @@
 // The Editor is displayed in the ABDesigner as a view is worked on.
 // The Editor allows a widget to be moved and placed on the canvas.
 //
-export default function FNAbviewgridEditor({ AB, ABViewEditorPlugin }) {
+export default function FNAbviewgridEditor({ ABViewEditorPlugin }) {
    // var L = UIClass.L();
-      // var L = ABViewContainer.L();
+   // var L = ABViewContainer.L();
 
-return class ABAbviewgridEditor extends ABViewEditorPlugin {
-
-static getPluginKey() {
+   return class ABAbviewgridEditor extends ABViewEditorPlugin {
+      static getPluginKey() {
          return this.key;
       }
 
-/**
+      /**
        * @method getPluginType
        * return the plugin type for this editor.
-       * plugin types are how our ClassManager knows how to store 
+       * plugin types are how our ClassManager knows how to store
        * the plugin.
        * @return {string} plugin type
        */
@@ -25,44 +24,39 @@ static getPluginKey() {
          // editor-view : will display in the editor panel of the ABDesigner
       }
 
+      static get key() {
+         return "grid";
+      }
 
+      constructor(view, base = "interface_editor_viewgrid") {
+         // base: {string} unique base id reference
 
+         super(view, base, {
+            label: "",
+         });
+         this.settings = view.settings;
+         // shortcut to reference the settings
 
-         static get key() {
-            return "grid";
-         }
+         this.base = base;
+      }
 
-         constructor(view, base = "interface_editor_viewgrid") {
-            // base: {string} unique base id reference
+      ui() {
+         return this.component.ui();
+      }
 
-            super(view, base, {
-               label: "",
-            });
-            this.settings = view.settings;
-            // shortcut to reference the settings
+      init(AB) {
+         this.AB = AB;
+         this.component.ignoreLocal = true;
+         return this.component.init(AB, 2);
+         // in our editor, we provide accessLv = 2
+      }
 
-            this.base = base;
-         }
+      detatch() {
+         this.component.detatch?.();
+      }
 
-         ui() {
-            return this.component.ui();
-         }
-
-         init(AB) {
-            this.AB = AB;
-            this.component.ignoreLocal = true;
-            return this.component.init(AB, 2);
-            // in our editor, we provide accessLv = 2
-         }
-
-         detatch() {
-            this.component.detatch?.();
-         }
-
-         onShow() {
-            this.component.onShow?.();
-         }
-      };
-   
-
+      onShow() {
+         this.component.onShow?.();
+      }
+   };
 }
