+++ src/plugins/web_view_grid/popupCountColumns.js
@@ -1,32 +1,27 @@
 /*
- * ui_work_object_workspace_popupSummaryColumns
+ * popupCountColumns
  *
- * Manage the Summary Columns popup.
+ * Manage the Count Columns popup.
  *
  */
-import UI_Class from "./ui_class";
+
+import UI_Class from "../../rootPages/Designer/ui_class";
 
 export default function (AB, ibase) {
-   ibase = ibase || "ui_work_object_workspace_popupSummaryColumns";
+   ibase = ibase || "popupCountColumns";
    // const uiConfig = AB.Config.uiSettings();
    const UIClass = UI_Class(AB);
    var L = UIClass.L();
 
-   class UI_Work_Object_Workspace_PopupSummaryColumns extends UIClass {
+   class UI_Work_Object_Workspace_PopupCountColumns extends UIClass {
       constructor(base) {
          super(base, {
             list: "",
          });
 
-         // make a reference to our Field CLASS definitions
-         this.ABFieldNumber = AB.Class.ABFieldManager.fieldByKey("number");
-         this.ABFieldFormula = AB.Class.ABFieldManager.fieldByKey("formula");
-         this.ABFieldCalculate =
-            AB.Class.ABFieldManager.fieldByKey("calculate");
-
-         this.SummaryFieldIds = [];
+         this.CountFieldIds = [];
          // {array}
-         // an array of the {ABFieldXXX.id}s that are used as summary fields.
+         // An array of the ABField.ids that are our count fileds.
       }
 
       // Our webix UI definition:
@@ -34,40 +29,40 @@ export default function (AB, ibase) {
          const ids = this.ids;
 
          return {
-            view: "popup",
             id: ids.component,
+            view: "popup",
             body: {
                rows: [
                   {
                      cols: [
                         {
                            view: "button",
-                           value: L("Select All"),
-                           on: {
-                              onItemClick: () => {
-                                 this.clickHideAll();
-                              },
-                           },
-                        },
-                        {
-                           view: "button",
                            css: "webix_primary",
-                           value: L("Unselect All"),
+                           value: L("Select All"),
                            on: {
                               onItemClick: () => {
                                  this.clickShowAll();
                               },
                            },
                         },
+                        {
+                           view: "button",
+                           value: L("Unselect All"),
+                           on: {
+                              onItemClick: () => {
+                                 this.clickHideAll();
+                              },
+                           },
+                        },
                      ],
                   },
                   {
-                     view: "list",
                      id: ids.list,
+                     view: "list",
                      maxHeight: 250,
                      select: false,
                      template:
-                        '<span style="min-width: 18px; display: inline-block;"><i class="fa ab-summary-field-icon"></i>&nbsp;</span> #label#',
+                        '<span style="min-width: 18px; display: inline-block;"><i class="fa ab-count-field-icon"></i>&nbsp;</span> #label#',
                      on: {
                         onItemClick: (id, e, node) => {
                            this.clickListItem(id, e, node);
@@ -89,43 +84,46 @@ export default function (AB, ibase) {
          this.AB = AB;
 
          webix.ui(this.ui());
-
          return Promise.resolve();
       }
 
+      onChange() {
+         this.emit("changed", this.CountFieldIds);
+      }
+
       /**
-       * @function clickHideAll
+       * @method clickHideAll
        * the user clicked the [hide all] option.  So hide all our fields.
        */
       clickHideAll() {
          var List = $$(this.ids.list);
 
          // pass an array is empty
-         this.SummaryFieldIds = [];
+         this.CountFieldIds = [];
 
          // hide all icons
          List.find({}).forEach((item) => {
             this.iconHide(item.id);
          });
 
-         this.emit("changed", this.SummaryFieldIds);
-         // _logic.callbacks.onChange(SummaryFieldIds);
+         this.onChange();
       }
 
       /**
-       * @function clickShowAll
+       * @method clickShowAll
        * the user clicked the [show all] option.  So show all our fields.
        */
       clickShowAll() {
          var List = $$(this.ids.list);
 
-         this.SummaryFieldIds = List.find({}).map((f) => f.id);
+         this.CountFieldIds = List.find({}).map((f) => f.id);
 
          // show all icons
          List.find({}).forEach((item) => {
             this.iconShow(item.id);
          });
-         this.emit("changed", this.SummaryFieldIds);
+
+         this.onChange();
       }
 
       /**
@@ -134,20 +132,21 @@ export default function (AB, ibase) {
        */
       clickListItem(fieldId) {
          // select
-         if (this.SummaryFieldIds.indexOf(fieldId) < 0) {
-            this.SummaryFieldIds.push(fieldId);
+         if (this.CountFieldIds.indexOf(fieldId) < 0) {
+            this.CountFieldIds.push(fieldId);
 
             this.iconShow(fieldId);
          }
          // unselect
          else {
-            this.SummaryFieldIds = this.SummaryFieldIds.filter(
+            this.CountFieldIds = this.CountFieldIds.filter(
                (fid) => fid != fieldId
             );
 
             this.iconHide(fieldId);
          }
-         this.emit("changed", this.SummaryFieldIds);
+
+         this.onChange();
       }
 
       /**
@@ -155,7 +154,7 @@ export default function (AB, ibase) {
        * return the current value of the Summary Columns settings.
        */
       getValue() {
-         return this.SummaryFieldIds;
+         return this.CountFieldIds;
       }
 
       /**
@@ -168,7 +167,7 @@ export default function (AB, ibase) {
          var $node = List.getItemNode(fieldId);
          if ($node) {
             $node
-               .querySelector(".ab-summary-field-icon")
+               .querySelector(".ab-count-field-icon")
                .classList.remove("fa-circle");
          }
       }
@@ -183,19 +182,10 @@ export default function (AB, ibase) {
          var $node = List.getItemNode(fieldId);
          if ($node) {
             $node
-               .querySelector(".ab-summary-field-icon")
+               .querySelector(".ab-count-field-icon")
                .classList.add("fa-circle");
          }
       }
-
-      /**
-       * @function objectLoad
-       * Ready the Popup according to the current object
-       * @param {ABObject} object  the currently selected object.
-       */
-      // objectLoad (object) {
-      //    CurrentObject = object;
-      // }
 
       /**
        * @function setValue
@@ -203,8 +193,7 @@ export default function (AB, ibase) {
        * @param {array} - an array contains field ids
        */
       setValue(fieldIds) {
-         this.SummaryFieldIds = fieldIds || [];
-
+         this.CountFieldIds = fieldIds || [];
          this.onShow();
       }
 
@@ -213,44 +202,19 @@ export default function (AB, ibase) {
        * Ready the Popup according to the current object each time it is shown (perhaps a field was created or delted)
        */
       onShow() {
-         const ids = this.ids;
-
          // refresh list
-         var numberFields = this.CurrentObject.fields(
-            (f) => f instanceof this.ABFieldNumber
-         ).map((f) => {
+         var allFields = this.CurrentObject.fields().map((f) => {
             return {
                id: f.id,
                label: f.label,
             };
          });
 
-         var calculateFields = this.CurrentObject.fields(
-            (f) => f instanceof this.ABFieldCalculate
-         ).map((f) => {
-            return {
-               id: f.id,
-               label: f.label,
-            };
-         });
-
-         var formulaFields = this.CurrentObject.fields(
-            (f) => f instanceof this.ABFieldFormula
-         ).map((f) => {
-            return {
-               id: f.id,
-               label: f.label,
-            };
-         });
-
-         var fieldOptions = numberFields.concat(calculateFields);
-         fieldOptions = fieldOptions.concat(formulaFields);
-
-         $$(ids.list).clearAll();
-         $$(ids.list).parse(fieldOptions);
+         $$(this.ids.list).clearAll();
+         $$(this.ids.list).parse(allFields);
 
          // update icons
-         this.SummaryFieldIds.forEach((fieldId) => {
+         this.CountFieldIds.forEach((fieldId) => {
             this.iconShow(fieldId);
          });
       }
@@ -265,6 +229,5 @@ export default function (AB, ibase) {
          $$(this.ids.component).show(...params);
       }
    }
-
-   return new UI_Work_Object_Workspace_PopupSummaryColumns(ibase);
+   return new UI_Work_Object_Workspace_PopupCountColumns(ibase);
 }
