+++ src/plugins/web_view_grid/popupSummaryColumns.js
@@ -1,27 +1,32 @@
 /*
- * ui_work_object_workspace_popupCountColumns
+ * popupSummaryColumns
  *
- * Manage the Count Columns popup.
+ * Manage the Summary Columns popup.
  *
  */
-
-import UI_Class from "./ui_class";
+import UI_Class from "../../rootPages/Designer/ui_class";
 
 export default function (AB, ibase) {
-   ibase = ibase || "ui_work_object_workspace_popupCountColumns";
+   ibase = ibase || "popupSummaryColumns";
    // const uiConfig = AB.Config.uiSettings();
    const UIClass = UI_Class(AB);
    var L = UIClass.L();
 
-   class UI_Work_Object_Workspace_PopupCountColumns extends UIClass {
+   class UI_Work_Object_Workspace_PopupSummaryColumns extends UIClass {
       constructor(base) {
          super(base, {
             list: "",
          });
 
-         this.CountFieldIds = [];
+         // make a reference to our Field CLASS definitions
+         this.ABFieldNumber = AB.Class.ABFieldManager.fieldByKey("number");
+         this.ABFieldFormula = AB.Class.ABFieldManager.fieldByKey("formula");
+         this.ABFieldCalculate =
+            AB.Class.ABFieldManager.fieldByKey("calculate");
+
+         this.SummaryFieldIds = [];
          // {array}
-         // An array of the ABField.ids that are our count fileds.
+         // an array of the {ABFieldXXX.id}s that are used as summary fields.
       }
 
       // Our webix UI definition:
@@ -29,40 +34,40 @@ export default function (AB, ibase) {
          const ids = this.ids;
 
          return {
-            id: ids.component,
             view: "popup",
+            id: ids.component,
             body: {
                rows: [
                   {
                      cols: [
                         {
                            view: "button",
-                           css: "webix_primary",
                            value: L("Select All"),
-                           on: {
-                              onItemClick: () => {
-                                 this.clickShowAll();
-                              },
-                           },
-                        },
-                        {
-                           view: "button",
-                           value: L("Unselect All"),
                            on: {
                               onItemClick: () => {
                                  this.clickHideAll();
                               },
                            },
                         },
+                        {
+                           view: "button",
+                           css: "webix_primary",
+                           value: L("Unselect All"),
+                           on: {
+                              onItemClick: () => {
+                                 this.clickShowAll();
+                              },
+                           },
+                        },
                      ],
                   },
                   {
-                     id: ids.list,
                      view: "list",
+                     id: ids.list,
                      maxHeight: 250,
                      select: false,
                      template:
-                        '<span style="min-width: 18px; display: inline-block;"><i class="fa ab-count-field-icon"></i>&nbsp;</span> #label#',
+                        '<span style="min-width: 18px; display: inline-block;"><i class="fa ab-summary-field-icon"></i>&nbsp;</span> #label#',
                      on: {
                         onItemClick: (id, e, node) => {
                            this.clickListItem(id, e, node);
@@ -84,46 +89,43 @@ export default function (AB, ibase) {
          this.AB = AB;
 
          webix.ui(this.ui());
+
          return Promise.resolve();
       }
 
-      onChange() {
-         this.emit("changed", this.CountFieldIds);
-      }
-
       /**
-       * @method clickHideAll
+       * @function clickHideAll
        * the user clicked the [hide all] option.  So hide all our fields.
        */
       clickHideAll() {
          var List = $$(this.ids.list);
 
          // pass an array is empty
-         this.CountFieldIds = [];
+         this.SummaryFieldIds = [];
 
          // hide all icons
          List.find({}).forEach((item) => {
             this.iconHide(item.id);
          });
 
-         this.onChange();
+         this.emit("changed", this.SummaryFieldIds);
+         // _logic.callbacks.onChange(SummaryFieldIds);
       }
 
       /**
-       * @method clickShowAll
+       * @function clickShowAll
        * the user clicked the [show all] option.  So show all our fields.
        */
       clickShowAll() {
          var List = $$(this.ids.list);
 
-         this.CountFieldIds = List.find({}).map((f) => f.id);
+         this.SummaryFieldIds = List.find({}).map((f) => f.id);
 
          // show all icons
          List.find({}).forEach((item) => {
             this.iconShow(item.id);
          });
-
-         this.onChange();
+         this.emit("changed", this.SummaryFieldIds);
       }
 
       /**
@@ -132,21 +134,20 @@ export default function (AB, ibase) {
        */
       clickListItem(fieldId) {
          // select
-         if (this.CountFieldIds.indexOf(fieldId) < 0) {
-            this.CountFieldIds.push(fieldId);
+         if (this.SummaryFieldIds.indexOf(fieldId) < 0) {
+            this.SummaryFieldIds.push(fieldId);
 
             this.iconShow(fieldId);
          }
          // unselect
          else {
-            this.CountFieldIds = this.CountFieldIds.filter(
+            this.SummaryFieldIds = this.SummaryFieldIds.filter(
                (fid) => fid != fieldId
             );
 
             this.iconHide(fieldId);
          }
-
-         this.onChange();
+         this.emit("changed", this.SummaryFieldIds);
       }
 
       /**
@@ -154,7 +155,7 @@ export default function (AB, ibase) {
        * return the current value of the Summary Columns settings.
        */
       getValue() {
-         return this.CountFieldIds;
+         return this.SummaryFieldIds;
       }
 
       /**
@@ -167,7 +168,7 @@ export default function (AB, ibase) {
          var $node = List.getItemNode(fieldId);
          if ($node) {
             $node
-               .querySelector(".ab-count-field-icon")
+               .querySelector(".ab-summary-field-icon")
                .classList.remove("fa-circle");
          }
       }
@@ -182,10 +183,19 @@ export default function (AB, ibase) {
          var $node = List.getItemNode(fieldId);
          if ($node) {
             $node
-               .querySelector(".ab-count-field-icon")
+               .querySelector(".ab-summary-field-icon")
                .classList.add("fa-circle");
          }
       }
+
+      /**
+       * @function objectLoad
+       * Ready the Popup according to the current object
+       * @param {ABObject} object  the currently selected object.
+       */
+      // objectLoad (object) {
+      //    CurrentObject = object;
+      // }
 
       /**
        * @function setValue
@@ -193,7 +203,8 @@ export default function (AB, ibase) {
        * @param {array} - an array contains field ids
        */
       setValue(fieldIds) {
-         this.CountFieldIds = fieldIds || [];
+         this.SummaryFieldIds = fieldIds || [];
+
          this.onShow();
       }
 
@@ -202,19 +213,44 @@ export default function (AB, ibase) {
        * Ready the Popup according to the current object each time it is shown (perhaps a field was created or delted)
        */
       onShow() {
+         const ids = this.ids;
+
          // refresh list
-         var allFields = this.CurrentObject.fields().map((f) => {
+         var numberFields = this.CurrentObject.fields(
+            (f) => f instanceof this.ABFieldNumber
+         ).map((f) => {
             return {
                id: f.id,
                label: f.label,
             };
          });
 
-         $$(this.ids.list).clearAll();
-         $$(this.ids.list).parse(allFields);
+         var calculateFields = this.CurrentObject.fields(
+            (f) => f instanceof this.ABFieldCalculate
+         ).map((f) => {
+            return {
+               id: f.id,
+               label: f.label,
+            };
+         });
+
+         var formulaFields = this.CurrentObject.fields(
+            (f) => f instanceof this.ABFieldFormula
+         ).map((f) => {
+            return {
+               id: f.id,
+               label: f.label,
+            };
+         });
+
+         var fieldOptions = numberFields.concat(calculateFields);
+         fieldOptions = fieldOptions.concat(formulaFields);
+
+         $$(ids.list).clearAll();
+         $$(ids.list).parse(fieldOptions);
 
          // update icons
-         this.CountFieldIds.forEach((fieldId) => {
+         this.SummaryFieldIds.forEach((fieldId) => {
             this.iconShow(fieldId);
          });
       }
@@ -229,5 +265,6 @@ export default function (AB, ibase) {
          $$(this.ids.component).show(...params);
       }
    }
-   return new UI_Work_Object_Workspace_PopupCountColumns(ibase);
+
+   return new UI_Work_Object_Workspace_PopupSummaryColumns(ibase);
 }
